move TransactionalMap.wrapIterator() -> MVSpartialIndex
