+++ h2/src/main/org/h2/mvstore/db/MVSpatialIndex.java
@@ -5,6 +5,7 @@
  */
 package org.h2.mvstore.db;
 
+import org.h2.mvstore.DataUtils;
 import org.h2.mvstore.MVMap;
 import static org.h2.util.geometry.GeometryUtils.MAX_X;
 import static org.h2.util.geometry.GeometryUtils.MAX_Y;
@@ -139,7 +140,7 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
         if (indexType.isUnique()) {
             // this will detect committed entries only
             RTreeCursor cursor = spatialMap.findContainedKeys(key);
-            Iterator<SpatialKey> it = map.wrapIterator(cursor, false);
+            Iterator<SpatialKey> it = new SpatialKeyIterator(map, cursor, false);
             while (it.hasNext()) {
                 SpatialKey k = it.next();
                 if (k.equalsIgnoringId(key)) {
@@ -155,7 +156,7 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
         if (indexType.isUnique()) {
             // check if there is another (uncommitted) entry
             RTreeCursor cursor = spatialMap.findContainedKeys(key);
-            Iterator<SpatialKey> it = map.wrapIterator(cursor, true);
+            Iterator<SpatialKey> it = new SpatialKeyIterator(map, cursor, true);
             while (it.hasNext()) {
                 SpatialKey k = it.next();
                 if (k.equalsIgnoringId(key)) {
@@ -207,7 +208,7 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
     private Cursor find(Session session) {
         Iterator<SpatialKey> cursor = spatialMap.keyIterator(null);
         TransactionMap<SpatialKey, Value> map = getMap(session);
-        Iterator<SpatialKey> it = map.wrapIterator(cursor, false);
+        Iterator<SpatialKey> it = new SpatialKeyIterator(map, cursor, false);
         return new MVStoreCursor(session, it, mvTable);
     }
 
@@ -221,7 +222,7 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
         Iterator<SpatialKey> cursor =
                 spatialMap.findIntersectingKeys(getKey(intersection));
         TransactionMap<SpatialKey, Value> map = getMap(session);
-        Iterator<SpatialKey> it = map.wrapIterator(cursor, false);
+        Iterator<SpatialKey> it = new SpatialKeyIterator(map, cursor, false);
         return new MVStoreCursor(session, it, mvTable);
     }
 
@@ -403,10 +404,11 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
         return dataMap.map;
     }
 
+
     /**
      * A cursor.
      */
-    public static class MVStoreCursor implements Cursor {
+    private static class MVStoreCursor implements Cursor {
 
         private final Session session;
         private final Iterator<SpatialKey> it;
@@ -415,7 +417,7 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
         private SearchRow searchRow;
         private Row row;
 
-        public MVStoreCursor(Session session, Iterator<SpatialKey> it, MVTable mvTable) {
+        MVStoreCursor(Session session, Iterator<SpatialKey> it, MVTable mvTable) {
             this.session = session;
             this.it = it;
             this.mvTable = mvTable;
@@ -465,6 +467,50 @@ public class MVSpatialIndex extends BaseIndex implements SpatialIndex, MVIndex {
             throw DbException.getUnsupportedException("previous");
         }
 
+    }
+
+    private static class SpatialKeyIterator implements Iterator<SpatialKey>
+    {
+        private final TransactionMap<SpatialKey, Value> map;
+        private final Iterator<SpatialKey> iterator;
+        private final boolean includeUncommitted;
+        private SpatialKey current;
+
+        SpatialKeyIterator(TransactionMap<SpatialKey, Value> map,
+                            Iterator<SpatialKey> iterator, boolean includeUncommitted) {
+            this.map = map;
+            this.iterator = iterator;
+            this.includeUncommitted = includeUncommitted;
+            fetchNext();
+        }
+
+        private void fetchNext() {
+            while (iterator.hasNext()) {
+                current = iterator.next();
+                if (includeUncommitted || map.containsKey(current)) {
+                    return;
+                }
+            }
+            current = null;
+        }
+
+        @Override
+        public boolean hasNext() {
+            return current != null;
+        }
+
+        @Override
+        public SpatialKey next() {
+            SpatialKey result = current;
+            fetchNext();
+            return result;
+        }
+
+        @Override
+        public void remove() {
+            throw DataUtils.newUnsupportedOperationException(
+                    "Removal is not supported");
+        }
     }
 
     /**
